+++ src/server/tools.rs
@@ -165,6 +165,24 @@ pub struct TypeDefinitionParams {
     pub character: u64,
 }
 
+/// Implementation parameters
+#[derive(Debug, serde::Deserialize, schemars::JsonSchema)]
+pub struct ImplementationParams {
+    /// Unique identifier for the target Neovim instance
+    pub connection_id: String,
+    /// Universal document identifier
+    // Supports both string and struct deserialization.
+    // Compatible with Claude Code when using subscription.
+    #[serde(deserialize_with = "string_or_struct")]
+    pub document: DocumentIdentifier,
+    /// Lsp client name
+    pub lsp_client_name: String,
+    /// Symbol position, line number starts from 0
+    pub line: u64,
+    /// Symbol position, character number starts from 0
+    pub character: u64,
+}
+
 /// Code action resolve parameters
 #[derive(Debug, serde::Deserialize, schemars::JsonSchema)]
 pub struct ResolveCodeActionParams {
@@ -498,6 +516,28 @@ impl NeovimMcpServer {
             .await?;
         Ok(CallToolResult::success(vec![Content::json(
             type_definition,
+        )?]))
+    }
+
+    #[tool(description = "Get LSP implementation")]
+    #[instrument(skip(self))]
+    pub async fn lsp_implementations(
+        &self,
+        Parameters(ImplementationParams {
+            connection_id,
+            document,
+            lsp_client_name,
+            line,
+            character,
+        }): Parameters<ImplementationParams>,
+    ) -> Result<CallToolResult, McpError> {
+        let client = self.get_connection(&connection_id)?;
+        let position = Position { line, character };
+        let implementation = client
+            .lsp_implementation(&lsp_client_name, document, position)
+            .await?;
+        Ok(CallToolResult::success(vec![Content::json(
+            implementation,
         )?]))
     }